+++ 10xdev/src/App.js
@@ -1,5 +1,5 @@
 import Navbar from './components/Navbar';
-import EdittableTable from './components/EdittableTable';
+import CodeTab from './components/CodeTab';
 import './App.css';
 import Tabs from "./components/Tabs";
 
@@ -7,7 +7,7 @@ function App() {
   return (
     <div>
         <Navbar/>
-        <EdittableTable/>
+        <CodeTab/>
     </div>
     );
 }
