+++ 10xdev/src/components/Navbar.js
@@ -8,8 +8,10 @@ function Navbar() {
         10xDEV.AI
       </div>
       <div className="search">
-          <input type="text" placeholder="What feature would you like to implement?" />
-          <button>Go!</button>
+          <input type="text" placeholder="  What feature would you like to implement?" />
+      </div>
+      <div className="button">
+        <button>Go!</button>
       </div>
     </nav>
   );
